backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# RAG stack
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-qdrant==0.2.0
qdrant-client==1.12.1
sentence-transformers==3.3.1
pypdf==5.1.0

# Data
motor==3.6.0
pymongo==4.9.2

# Auth
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
